requirements.txt
Django==4.2.7
djangorestframework==3.15.2
drf-spectacular==0.27.2
flake8==7.1.1
python-decouple==3.8
pytest==8.3.3
pytest-django==4.9.0